+++ app/Http/Controllers/BoardController.php
@@ -2,11 +2,11 @@
 
 namespace App\Http\Controllers;
 
-use App\Chapter;
-use App\FinancialReport;
+use App\Models\Chapter;
+use App\Models\FinancialReport;
 use App\Mail\ChapersUpdateListAdmin;
 use App\Mail\ChapersUpdatePrimaryCoor;
-use App\User;
+use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
