+++ app/Http/Controllers/CoordinatorController.php
@@ -2,10 +2,10 @@
 
 namespace App\Http\Controllers;
 
-use App\Coordinator;
-use App\FinancialReport;
+use App\Models\Coordinator;
+use App\Models\FinancialReport;
 use App\Mail\ChapersUpdatePrimaryCoor;
-use App\User;
+use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
