+++ app/Http/Controllers/ReportController.php
@@ -2,9 +2,9 @@
 
 namespace App\Http\Controllers;
 
-use App\Chapter;
-use App\Coordinator;
-use App\User;
+use App\Models\Chapter;
+use App\Models\Coordinator;
+use App\Models\User;
 use Exception;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
