+++ app/Models/CoordinatorDetails.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 
-class Chapter extends Model
+class CoordinatorDetails extends Model
 {
     //
     public $timestamps = false;
@@ -14,4 +14,6 @@ class Chapter extends Model
         'name',
 
     ];
+
+    protected $table = 'coordinator_details';
 }
