+++ resources/views/coordinators/dashboard.blade.php
@@ -27,7 +27,7 @@
     <form method="POST" action='{{ route("coordinator.updatedashboard",$coordinatorDetails[0]->coordinator_id) }}' id="update-dashboard">
     @csrf
         <?php 
-        use App\User;
+        use App\Models\User;
         $corDetails = User::find(Auth::user()->id)->CoordinatorDetails;
         $corId = $corDetails['coordinator_id'];
         $positionid = $corDetails['position_id'];
